+++ Runtime/Config/FlockInitConfig.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Flock.Analytics;
 using Flock.Http;
 
 namespace Flock.Config
@@ -11,13 +12,15 @@ namespace Flock.Config
         public string GameVersionId { get; set; }
         public bool EnableDebugLogs { get; set; }
         public RetryPolicy RetryPolicy { get; set; }
-
+        public FlockAnalyticsConfig Analytics { get; set; }
+        private Dictionary<string, string> _headers;
         public FlockInitConfig(
             string apiUrl,
             string apiKey,
             string gameId,
             string gameVersionId,
             bool enableDebugLogs = false,
+            FlockAnalyticsConfig analytics = null,
             RetryPolicy retryPolicy = null)
         {
             ApiUrl = apiUrl;
@@ -25,21 +28,16 @@ namespace Flock.Config
             GameId = gameId;
             GameVersionId = gameVersionId;
             EnableDebugLogs = enableDebugLogs;
+            Analytics = analytics;
             RetryPolicy = retryPolicy ?? new RetryPolicy();
         }
 
         public Dictionary<string, string> GetBaseHeaders()
         {
-            var headers = new Dictionary<string, string>
-            {
-                { "X-Flock-API-Key", ApiKey }
-            };
-
-            if (!string.IsNullOrEmpty(GameVersionId))
-                headers["X-Game-Version-ID"] = GameVersionId;
-
-            return headers;
+            _headers ??= new Dictionary<string, string>();
+            _headers.TryAdd("X-Flock-API-Key", ApiKey);
+            _headers.TryAdd("X-Game-Version-ID", GameVersionId);
+            return _headers;
         }
-
     }
 }