+++ Runtime/Providers/FlockSchemaProvider.cs
@@ -8,19 +8,17 @@ using Flock.Interfaces;
 
 namespace Flock.Config
 {
+  
     public class FlockSchemaProvider : FlockProviderBase, ISchemaProvider
     {
         public FlockSchemaProvider(FlockClient client) : base(client) { }
 
-        public async Task<List<GameConfigSchema>> GetAllSchemasAsync(string tag = null, CancellationToken cancellationToken = default)
+        public async Task<List<GameConfigSchema>> GetAllSchemasAsync(SchemaTag tag , CancellationToken cancellationToken = default)
         {
             return await ExecuteAsync(async () =>
             {
                 var url = new StringBuilder().Append(Client.GetApiUrl()).Append("/v1/game_config");
-
-                if (!string.IsNullOrEmpty(tag))
-                    url.Append("?tag=").Append(tag);
-
+                url.Append("?tag=").Append(tag.ToString());
                 var response = await FlockHttpClient.GetAsync<GenericResponse<List<GameConfigSchema>>>(
                     url.ToString(), Client.GetBaseHeaders(), cancellationToken);
                 ValidateResponse(response);
@@ -28,15 +26,12 @@ namespace Flock.Config
             }, "Fetch config schemas", cancellationToken);
         }
 
-        public async Task<List<GameConfigSchema>> GetSchemasByVersionAsync(string tag = null, CancellationToken cancellationToken = default)
+        public async Task<List<GameConfigSchema>> GetSchemasByVersionAsync(SchemaTag tag, CancellationToken cancellationToken = default)
         {
             return await ExecuteAsync(async () =>
             {
                 var url = new StringBuilder().Append(Client.GetApiUrl()).Append("/v1/game_config/version");
-
-                if (!string.IsNullOrEmpty(tag))
-                    url.Append("?tag=").Append(tag);
-
+                url.Append("?tag=").Append(tag.ToString());
                 var response = await FlockHttpClient.GetAsync<GenericResponse<List<GameConfigSchema>>>(
                     url.ToString(), Client.GetBaseHeaders(), cancellationToken);
                 ValidateResponse(response);